requirements.txt
metatrader5==5.0.5430
matplotlib==3.9.2
python-dotenv==1.0.1
